user_service/requirements.txt
aio-pika==9.4.3
aiormq==6.8.0
aiosmtplib==3.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
pamqp==3.3.0
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.9.4

